backend/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
inexactsearch==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pyspellchecker==0.8.1
silpa_common==0.3
soundex==1.1.3
Werkzeug==3.0.1
